# perf: avoid importing npm-registry-fetch for replace-info