Instead of importing the entire module, just copy the function from the `npm-registry-fetch`.

Benchmark:

```bash
hyperfine --warmup 3 "node ./bin/npm-cli.js run echo"
```

Before:

```
Benchmark 1: node ./bin/npm-cli.js run echo
  Time (mean ± σ):     219.8 ms ±   4.0 ms    [User: 224.8 ms, System: 47.2 ms]
  Range (min … max):   213.2 ms … 225.9 ms    13 runs
```

After:

```
Benchmark 1: node ./bin/npm-cli.js run echo
  Time (mean ± σ):     180.0 ms ±   5.9 ms    [User: 178.7 ms, System: 34.7 ms]
  Range (min … max):   165.5 ms … 191.4 ms    16 runs
```