@lukekarrys 2024-03-26
We are going to publish the replace info util as its own module since we need a version of it elsewhere as well.

I believe that should have a similar performance impact. But I don't think this PR will land as is. Thank you for this though! I will keep the open until we have an open PR implementing it as a separate package so we can benchmark that also.